Thuan_Code bảng Hóa Đơn

+++ QuanLySanPham/src/main/java/Views/FarmeQLSanPham.java
@@ -144,7 +144,7 @@ public class FarmeQLSanPham extends javax.swing.JFrame {
     }
 
     SanPhamModel getdata() {
-        return new SanPhamModel(txt_maSanPham.getText(), txt_tenSanPham.getText(), Integer.parseInt(txt_soLuong.getText()), Double.parseDouble(txt_giiaban.getText()), tar_mota.getText(), cbc_trangThai.getSelectedItem().equals("Đang Bán"),
+        return new SanPhamModel(txt_maSanPham.getText(), txt_tenSanPham.getText(), Integer.parseInt(txt_soLuong.getText()), Integer.parseInt(txt_giiaban.getText()), tar_mota.getText(), cbc_trangThai.getSelectedItem().equals("Đang Bán"),
                 _ChatLieuService.getProducts().get(cbc_chatLieu.getSelectedIndex()),
                 _KieuDangService.getproduct().get(cbc_kieuDang.getSelectedIndex()),
                 _DanhMucSanPhamService.getListFromDB().get(cbc_danhMucSanPham.getSelectedIndex()),
